+++ @types/react-textfit.d.ts
@@ -1,4 +1,4 @@
 declare module 'react-textfit' {
-  let TextFit: (props: any) => JSX.Element
-  export default TextFit
+  let Textfit: (props: any) => JSX.Element
+  export {Textfit}
 }
